twitter_bot/requirements.txt
feedparser==6.0.11
python-telegram-bot==20.7
google-generativeai==0.3.2
pytz==2023.3
python-dotenv==1.0.0